% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{CalculateSilhouetteInformation}
\alias{CalculateSilhouetteInformation}
\alias{CalculateSilhouetteInformation.SingleCellExperiment}
\alias{CalculateSilhouetteInformation,SingleCellExperiment-method}
\title{Estimating optimal K using silhouette}
\usage{
CalculateSilhouetteInformation.SingleCellExperiment(object, K.start, K.end)

\S4method{CalculateSilhouetteInformation}{SingleCellExperiment}(object, K.start = 2, K.end = 50)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{K.start}{a numeric for the smallest
K value to be tested. Default is \code{2}.}

\item{K.end}{a numeric for the largest
K value to be tested. Default is \code{50}.}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
The function estimates the optimal number of clusters K from the dendrogram
of the hierarhical clustering using the silhouette method.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce,K.start=2, K.end=50)
}
\keyword{clustering}
\keyword{hierarchical}
\keyword{ward}
