% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{DownOverSampling}
\alias{DownOverSampling}
\title{Down- and oversample data}
\usage{
DownOverSampling(x, n = 50)
}
\arguments{
\item{x}{A character or numeric vector of data to down-and oversample.}

\item{n}{How many cells to include per cluster.}
}
\value{
a list containing the output of the LiblineaR prediction
}
\description{
The function implements a script down- and oversamples data to
include n cells.
}
\keyword{downsampling}
\keyword{oversampling}
