% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{GeneHeatmap}
\alias{GeneHeatmap}
\alias{GeneHeatmap.SingleCellExperiment}
\alias{GeneHeatmap,SingleCellExperiment-method}
\title{Heatmap visualization of the gene markers identified by FindAllGeneMarkers}
\usage{
GeneHeatmap.SingleCellExperiment(object, clustering.type, gene.markers)

\S4method{GeneHeatmap}{SingleCellExperiment}(object, clustering.type = "manual", gene.markers = NULL)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{clustering.type}{"manual" or "optimal". "manual" refers to the
clustering formed using the "SelectKClusters" function and "optimal"
to the clustering using the "CalculateSilhouetteInformation" function.
Default is "manual".}

\item{gene.markers}{a data frame of the gene markers generated
by FindAllGeneMarkers function. To accelerate the drawing, filtering
the dataframe by selecting e.g. top 10 genes is recommended.}
}
\value{
nothing
}
\description{
The GeneHeatmap function enables drawing a heatmap of the gene markers
identified by FindAllGeneMarkers, where the cell are grouped
by the clustering.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce, K.start = 2, K.end = 50)
## Draw silhouette curve
GeneHeatmap(sce)

}
\keyword{gene}
\keyword{grouped}
\keyword{heatmap}
