% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\name{HierarchicalClustering}
\alias{HierarchicalClustering}
\title{Hierarchical clustering using the Ward's method}
\usage{
\S4method{HierarchicalClustering}{iloreg}(iloreg.object = NULL)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}
}
\value{
iloreg object
}
\description{
Perform Hierarchical clustering using the Ward's method.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- HierarchicalClustering(iloreg_object)
}
\keyword{clustering}
\keyword{hierarchical}
\keyword{ward}
