% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{package}
\name{ILoReg-package}
\alias{ILoReg-package}
\alias{ILoReg}
\title{Iterative Logistic Regression (ILoReg) consensus clustering}
\description{
R package that enables supervised learning -based clustering through L1-regularized (LASSO) logistic regression.
Consensus clustering is performed by running ILoReg multiple times. PCA is used to reduce dimensionality of the
consensus probability matrix. The Ward's method is used to perform the final clustering.
The silhouette method can be used to choose the optimal number of clusters (K). Moreover, t-SNE or UMAP can be performed
onto the PCA-transformed data to visualize the results.
}
\author{
Johannes Smolander <pajosm@utu.fi>, Mikko Ven?l?inen, Sini Junttila, Laura L. Elo
}
