% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Classification using logistic regression from LiblineaR R package}
\usage{
LogisticRegression(training.sparse.matrix = NULL,
  training.ident = NULL, print.info = TRUE, cost = 0.3,
  type = "L1", test.sparse.matrix = NULL,
  downsampling.fraction = 0.3)
}
\arguments{
\item{training.sparse.matrix}{A sparse matrix (dgCMatrix) containing training sample's gene expression data with genes in rows and cells in columns. (default NULL)}

\item{training.ident}{A named factor containing sample's cluster labels for each cell in training.sparse.matrix. (default NULL)}

\item{print.info}{A logical denoting if information about the run (e.g. running time) should be printed. (default TRUE)}

\item{cost}{Cost of constraints (C). Recommendation is in range [0.2,0.5] for LASSO.  (default 0.3)}

\item{type}{"L1" for LASSO and "L2" for Ridge.  (default "L1")}

\item{test.sparse.matrix}{A sparse matrix (dgCMatrix) containing test sample's gene expression data with genes in rows and cells in columns. (default NULL)}

\item{downsampling.fraction}{A numeric that defines how many cells per cluster should be down- and oversampled (d in N/k*d), when stratified.downsampling=FALSE,  or what fraction should be downsampled in the stratified approach ,stratified.downsampling=TRUE. (default 0.3)}

\item{stratified.downsampling}{A logical denoting if stratified downsampling approach should be used in generation of the training data: (default FALSE).}
}
\value{
predictions for test data
}
\description{
Enables training of a logistic regression model using LiblineaR package.
}
\examples{
a <- c(0,1,2)
}
\keyword{LiblineaR}
\keyword{logistic}
\keyword{regression}
\keyword{stratified}
