% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{pbmc3k_1000}
\alias{pbmc3k_1000}
\title{1000 cells downsampled from the pbmc3k dataset.}
\format{raw_data and data, both dgCMatrix objects}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression}
}
\usage{
data(pbmc3k_1000)
}
\description{
The preprocessing was done using Cell Ranger v2.2.0 and
the GRCh37.p13 human reference genome.
}
\examples{
data(pbmc3k_1000)
}
\keyword{datasets}
