% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{PCAElbowPlot}
\alias{PCAElbowPlot}
\alias{PCAElbowPlot.SingleCellExperiment}
\alias{PCAElbowPlot,SingleCellExperiment-method}
\title{Elbow plot of the standard deviations of the principal components}
\usage{
PCAElbowPlot.SingleCellExperiment(object, return.plot)

\S4method{PCAElbowPlot}{SingleCellExperiment}(object, return.plot = FALSE)
}
\arguments{
\item{object}{object of class 'iloreg'}

\item{return.plot}{logical indicating if the ggplot2 object
should be returned (default FALSE)}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
Draw an elbow plot of the standard deviations of the principal components
to deduce an appropriate value for p.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
##
sce <- RunPCA(sce,p = 50,scale = FALSE)
PCAElbowPlot(sce,return.plot=FALSE)

}
\keyword{PCA}
\keyword{elbow}
\keyword{plot}
