% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RenameAllClusters}
\alias{RenameAllClusters}
\alias{RenameAllClusters.SingleCellExperiment}
\alias{RenameAllClusters,SingleCellExperiment-method}
\title{Renaming all clusters at once}
\usage{
RenameAllClusters.SingleCellExperiment(object, new.cluster.names)

\S4method{RenameAllClusters}{SingleCellExperiment}(object, new.cluster.names = "")
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{new.cluster.names}{object of class 'iloreg'}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
RenameAllClusters function enables renaming all cluster at once.
}
\keyword{all}
\keyword{clusters}
\keyword{rename}
