% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunICP.R
\name{RunICP}
\alias{RunICP}
\title{Iterative logistic regression (ILoReg) clustering}
\usage{
RunICP(
  normalized.data = NULL,
  k = 15,
  d = 0.3,
  r = 5,
  C = 5,
  type = "L1",
  max.number.of.iterations = 100
)
}
\arguments{
\item{normalized.data}{A sparse matrix (dgCMatrix) containing gene expression with genes in rows and cells in columns. (default NULL)}

\item{k}{A positive integer greater or equal to 2, which denotes the number of clusters in iterative logistic regression. If stratified.downsampling=TRUE, the final k is likely to be smaller. (default 15)}

\item{d}{A numeric that defines how many cells per cluster should be down- and oversampled (d in N/k*d), when stratified.downsampling=FALSE,  or what fraction should be downsampled in the stratified approach ,stratified.downsampling=TRUE. (default 0.3)}

\item{r}{A positive integer that denotes the maximum number of reiterations performed until the algorithm ends. (default 5)}

\item{C}{Cost of constraints (C). Recommended C is in range [0.2,0.5] for LASSO.  (default 0.3)}

\item{type}{"L1" for LASSO and "L2" for Ridge.  (default "L1")}

\item{max.number.of.iterations}{A positive integer that denotes the maximum number of iterations performed until the algorithm ends. (default 200)}
}
\value{
return the probability matrix, and clutering comparison measures if return.metrics=TRUE or UMAP layouts from each epoch if perform.umap=TRUE
}
\description{
Enables iterative logistic regression (ILoReg) clustering.
}
\examples{
a <- c(0,1,2)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
