% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{RunILoRegConsensus}
\alias{RunILoRegConsensus}
\title{Iterative logistic regression (ILoReg) consensus clustering}
\usage{
\S4method{RunILoRegConsensus}{iloreg}(iloreg.object = NULL, k = 15,
  d = 0.3, L = 100, r = 5, C = 0.5, type = "L1",
  max.number.of.iterations = 200, threads = 0)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{k}{A positive integer greater or equal to 2, which denotes the number of clusters in iterative logistic regression. If stratified.downsampling=TRUE, the final k is likely to be smaller. (default 15)}

\item{d}{A numeric that defines how many cells per cluster should be down- and oversampled (d in N/k*d), when stratified.downsampling=FALSE,  or what fraction should be downsampled in the stratified approach ,stratified.downsampling=TRUE. (default 0.3)}

\item{L}{A positive integer that specifies how many ILoReg runs in the consensus method are ran. (default 50)}

\item{r}{A positive integer that denotes the maximum number of reiterations performed until the algorithm ends. (default 5)}

\item{C}{Cost of constraints (C). Recommendation is in range [0.2,0.5] for LASSO.  (default 0.3)}

\item{type}{"L1" for LASSO and "L2" for Ridge.  (default "L1")}

\item{max.number.of.iterations}{A positive integer that denotes the maximum number of iterations performed until the algorithm ends. (default 200)}

\item{threads}{A positive integer that specifies how many logical processors (threads) should be used. Use threads=1 to disable parallelism and threads=0 to use all available core minus one. (default 0)}
}
\value{
iloreg object
}
\description{
Enables iterative logistic regression (ILoReg) consensus clustering with parallel computation.
}
\details{
populates a Boolean matrix with the same dimension as nData.
The value is \code{TRUE} for an entry if it
is a dropout candidate; otherwise the value is \code{FALSE}.
}
\examples{
a <- c(0,1,2)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
