% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{RunParallelICP}
\alias{RunParallelICP}
\title{ICP consensus method}
\usage{
\S4method{RunParallelICP}{iloreg}(iloreg.object = NULL, k = 15,
  d = 0.3, L = 200, r = 5, C = 0.3, type = "L1",
  max.number.of.iterations = 200, threads = 0, seed = 1)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{k}{A positive integer greater or equal to 2, which denotes the number of clusters in iterative logistic regression. (default 15)}

\item{d}{A numeric greater than 0 and smaller than 1 that determines how many cells 'n' are down- or oversampled from each cluster into the training data. (d in n=N/k*d), where N is the total number of cells, k is the number of clusters in ICP. (default 0.3)}

\item{L}{A positive integer greater than 1. Number of ICP runs. Contraining L has not been studied.  (Default 200)}

\item{r}{A positive integer that denotes the maximum number of reiterations performed until the ICP algorithm stops. (default 5). Increasing recommended with a significantly larger sample size (tens of thousands).}

\item{C}{A positive real number, the cost of constraints violation in the L1-regularized logistic regression model from the LIBLINEAR library. (default 0.3). Decreasing leads to more stringent feature selection, i.e. less genes are selected that are used to build the classifier.}

\item{type}{"L1" or "L2". L2-regularization was not investigated in the manuscript. (default "L1")}

\item{max.number.of.iterations}{A positive integer that denotes the maximum number of iterations performed until the algorithm ends. (default 200)}

\item{threads}{A positive integer that specifies how many logical processors (threads) should be used. Use threads=1 to disable parallelism and threads=0 to use all available core minus one. (default 0)}

\item{seed}{A positive integer that specifies the random seed to be set before randomly generating random seeds for each parallel task. (default 1)}
}
\value{
iloreg object
}
\description{
Enables running L ICP runs in parallel.
}
\examples{
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
