% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunTSNE}
\alias{RunTSNE}
\alias{RunTSNE.SingleCellExperiment}
\alias{RunTSNE,SingleCellExperiment-method}
\title{Barnes-Hut implementation of t-Distributed Stochastic
Neighbor Embedding (t-SNE)}
\usage{
RunTSNE.SingleCellExperiment(object, perplexity)

\S4method{RunTSNE}{SingleCellExperiment}(object, perplexity = 30)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{perplexity}{perplexity of t-SNE}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Run nonlinear dimensionality reduction using t-SNE with the
PCA-transformed consensus matrix as input.
}
\keyword{Barnes-Hut}
\keyword{Embedding}
\keyword{Neighbor}
\keyword{Stochastic}
\keyword{implementation}
\keyword{of}
\keyword{t-Distributed}
\keyword{t-SNE}
