% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunUMAP}
\alias{RunUMAP}
\alias{RunUMAP.SingleCellExperiment}
\alias{RunUMAP,SingleCellExperiment-method}
\title{Uniform Manifold Approximation and Projection (UMAP)}
\usage{
RunUMAP.SingleCellExperiment(object)

\S4method{RunUMAP}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Run nonlinear dimensionality reduction using UMAP with
the PCA-transformed consensus matrix as input.
}
\keyword{Approximation}
\keyword{Manifold}
\keyword{Projection}
\keyword{UMAP}
\keyword{Uniform}
\keyword{and}
