% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{SelectKClusters}
\alias{SelectKClusters}
\alias{SelectKClusters.SingleCellExperiment}
\alias{SelectKClusters,SingleCellExperiment-method}
\title{Selecting K clusters from hierarchical clustering}
\usage{
SelectKClusters.SingleCellExperiment(object, K)

\S4method{SelectKClusters}{SingleCellExperiment}(object, K = NULL)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{K}{a positive integer denoting how many clusters to select}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Selects K clusters from the dendrogram.
}
\keyword{clusters}
\keyword{select}
