% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VisualizeQC}
\alias{VisualizeQC}
\alias{VisualizeQC.SingleCellExperiment}
\alias{VisualizeQC,SingleCellExperiment-method}
\title{Visualize the terminal projection accuracy, number of
epochs/iterations and the average pairwise ARI values.}
\usage{
VisualizeQC.SingleCellExperiment(object, return.plot)

\S4method{VisualizeQC}{SingleCellExperiment}(object, return.plot = FALSE)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{return.plot}{logical indicating if the ggplot2 object
should be returned. Default is \code{FALSE}.}
}
\value{
ggplot2 object if \code{return.plot=TRUE}
}
\description{
Draw violin plots of the terminal projection accuracy,
the number of epochs/iterations and the average pairwise ARI values.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)

}
\keyword{ARI}
\keyword{QC}
\keyword{accuracy}
\keyword{average}
\keyword{control}
\keyword{pairwise}
\keyword{projection}
\keyword{quality}
\keyword{terminal}
