% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VlnPlot}
\alias{VlnPlot}
\alias{VlnPlot.SingleCellExperiment}
\alias{VlnPlot,SingleCellExperiment-method}
\title{Gene expression visualization using violin plots}
\usage{
VlnPlot.SingleCellExperiment(object, clustering.type, genes, return.plot)

\S4method{VlnPlot}{SingleCellExperiment}(object, clustering.type = "manual", genes = NULL, return.plot = FALSE)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{clustering.type}{"manual" or "optimal". "manual"
refers to the clustering formed using the "SelectKClusters" function
and "optimal" to the clustering formed using the
"CalculateSilhouetteInformation" function. Default is "manual".}

\item{genes}{a character vector denoting the gene names that are visualized}

\item{return.plot}{return.plot whether to return the ggplot2 object
or just draw it. Default is \code{FALSE}.}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
The VlnPlot function enables visualizing expression levels of a gene,
or multiple genes, across clusters using Violin plots.
}
\keyword{plot}
\keyword{violin}
