% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{AnnotationScatterPlot}
\alias{AnnotationScatterPlot}
\alias{AnnotationScatterPlot.SingleCellExperiment}
\alias{AnnotationScatterPlot,SingleCellExperiment-method}
\title{Visualiation of a custom annotation over nonlinear
dimensionality reduction}
\usage{
AnnotationScatterPlot.SingleCellExperiment(
  object,
  annotation,
  return.plot,
  dim.reduction.type,
  point.size,
  show.legend
)

\S4method{AnnotationScatterPlot}{SingleCellExperiment}(
  object,
  annotation = NULL,
  return.plot = FALSE,
  dim.reduction.type = "",
  point.size = 0.7,
  show.legend = FALSE
)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{annotation}{a character vector, factor or numeric for the class labels.}

\item{return.plot}{return.plot whether to return the ggplot2 object or
just draw it. Default is \code{FALSE}.}

\item{dim.reduction.type}{"tsne" or "umap". Default is \code{tsne}.}

\item{point.size}{point size. Default is \code{0.7}.}

\item{show.legend}{a logical denoting whether to show the legend on the right
side of the plot. Default is \code{TRUE}.}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
The AnnotationScatterPlot enables visualizing arbitrary class labels
over the nonliner dimensionality reduction, e.g. t-SNE or UMAP.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce, K.start = 2, K.end = 50)
## Draw silhouette curve
AnnotationScatterPlot(sce)

}
\keyword{annotation}
\keyword{custom}
\keyword{dimensionality}
\keyword{nonlinear}
\keyword{reduction}
\keyword{t-sne}
\keyword{umap}
\keyword{visualization}
