% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{AnnotationScatterPlot}
\alias{AnnotationScatterPlot}
\title{Visualiation of a custom annotation over nonlinear dimensionality reduction}
\usage{
\S4method{AnnotationScatterPlot}{iloreg}(iloreg.object = NULL,
  annotation = c(), return.plot = FALSE, dim.reduction.type = "",
  point.size = 0.7)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{annotation}{a character vector, factor or numeric for the class labels.}

\item{return.plot}{return.plot whether to return the ggplot2 object or just draw it (default \code{FALSE})}

\item{dim.reduction.type}{"tsne" or "umap" (default "tsne")}

\item{point.size}{point size (default 0.7)}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
The AnnotationScatterPlot enables visualizing arbitrary class labels over the nonliner dimensionality reduction, e.g. t-SNE or UMAP.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- RunTSNE(iloreg_object,perplexity=30)
AnnotationScatterPlot(iloreg_object,annotation=c(),point.size=0.5,dim.reduction.type="tsne")
}
\keyword{annotation}
\keyword{custom}
\keyword{dimensionality}
\keyword{nonlinear}
\keyword{reduction}
\keyword{t-sne}
\keyword{umap}
\keyword{visualization}
