% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{ClusteringScatterPlot}
\alias{ClusteringScatterPlot}
\title{Visualize the clustering over nonliner dimensionality reduction}
\usage{
\S4method{ClusteringScatterPlot}{iloreg}(iloreg.object = NULL,
  clustering.type = "manual", return.plot = FALSE,
  dim.reduction.type = "", point.size = 0.7, title = "",
  show.legend = TRUE)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{clustering.type}{"manual" or "optimal". "manual" refers to the clustering formed using the "SelectKClusters" function and "optimal" to the clustering formed using the "CalculateSilhouetteInformation" function. (default "manual")}

\item{return.plot}{object of class 'iloreg' (default \code{FALSE})}

\item{dim.reduction.type}{"tsne" or "umap" (default "tsne")}

\item{point.size}{point size (default 0.7)}

\item{title}{text to write above the plot}

\item{show.legend}{whether to show the legend on the right side of the plot}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
ClusteringScatterPlot function enables visualizing the clustering over nonliner dimensionality reduction (t-SNE or UMAP).
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- RunTSNE(iloreg_object,perplexity=30)
iloreg_object <- HierarchicalClustering(iloreg_object)
iloreg_object <- SelectKClusters(iloreg_object,K=15)
ClusteringScatterPlot(iloreg_object,clustering.type="manual",dim.reduction.type = "tsne",point.size=0.7)
iloreg_object <- CalculateSilhouetteInformation(iloreg_object,k.range = 2:50)
ClusteringScatterPlot(iloreg_object,clustering.type="optimal",dim.reduction.type = "tsne",point.size=0.7)
iloreg_object <- RunUMAP(iloreg_object,perplexity=30)
ClusteringScatterPlot(iloreg_object,clustering.type="manual",dim.reduction.type = "umap",point.size=0.7)
ClusteringScatterPlot(iloreg_object,clustering.type="optimal",dim.reduction.type = "umap",point.size=0.7)
}
\keyword{clustering}
\keyword{dimensionality}
\keyword{nonlinear}
\keyword{plot}
\keyword{reduction}
\keyword{scatter}
