% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\name{CreateILoRegObject}
\alias{CreateILoRegObject}
\title{Create an ILoReg object}
\usage{
CreateILoRegObject(normalized.data)
}
\arguments{
\item{normalized.data}{a gene expression matrix with genes in rows and cells in columns...}
}
\description{
Create an ILoReg object
}
\examples{
## Generate simulated single-cell RNA-Seq tags.
library(Seurat)
raw_data <- Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- LogNormalize(raw_data)
iloreg <- CreateILoRegObject(normalized.data=data)

}
