% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{FindAllGeneMarkers}
\alias{FindAllGeneMarkers}
\title{Iterative logistic regression (ILoReg) consensus clustering}
\usage{
\S4method{FindAllGeneMarkers}{iloreg}(iloreg.object = NULL,
  clustering.type = "", test = "wilcox", logfc.threshold = 0.25,
  min.pct = 0.1, min.diff.pct = NULL, min.cells.cluster = 3,
  max.cells.per.cluster = NULL, random.seed = 1, pseudocount.use = 1,
  return.thresh = 0.01, only.pos = FALSE)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{clustering.type}{object of class 'iloreg'}

\item{test}{object of class 'iloreg'}

\item{logfc.threshold}{object of class 'iloreg'}

\item{min.pct}{object of class 'iloreg'}

\item{min.diff.pct}{object of class 'iloreg'}

\item{min.cells.cluster}{object of class 'iloreg'}

\item{max.cells.per.cluster}{object of class 'iloreg'}

\item{random.seed}{object of class 'iloreg'}

\item{pseudocount.use}{object of class 'iloreg'}

\item{return.thresh}{object of class 'iloreg'}

\item{only.pos}{object of class 'iloreg'}
}
\value{
iloreg Object
}
\description{
Plot histogram of the ARIs.
}
\details{
populates a Boolean matrix with the same dimension as nData.
The value is \code{TRUE} for an entry if it
is a dropout candidate; otherwise the value is \code{FALSE}.
}
\examples{
a <- c(0,1,2)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
