% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{GeneDropoutRatePlot}
\alias{GeneDropoutRatePlot}
\alias{GeneDropoutRatePlot.SingleCellExperiment}
\alias{GeneDropoutRatePlot,SingleCellExperiment-method}
\title{Dropout rate versus average non-zero expression}
\usage{
GeneDropoutRatePlot.SingleCellExperiment(
  object,
  genes,
  return.plot,
  clusters,
  clustering.type
)

\S4method{GeneDropoutRatePlot}{SingleCellExperiment}(
  object,
  genes = NULL,
  return.plot = FALSE,
  clusters = NULL,
  clustering.type = "manual"
)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{genes}{a character vector of the genes to visualize over
the dropout rate curve}

\item{return.plot}{return.plot whether to return the ggplot2 object or
just draw it. Default is \code{FALSE}.}

\item{clusters}{use data from these clusters only.
Default is \code{NULL}.}

\item{clustering.type}{"manual" or "optimal". "manual" refers to the
clustering formed using the "SelectKClusters" function and "optimal"
to the clustering using the "CalculateSilhouetteInformation" function.
Default is \code{manual}.}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
The GeneDropoutRatePlot function enables visualizion of dropout rates
(fraction of cells expressing a gene) against the average non-zero
expression values. This function can aid the user to determinine if
a gene is differentially expressed or not, since differentially expressed
genes are likely to be above the expected curve. Cells can be filtered based
on clustering before calculating the values.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce, K.start = 2, K.end = 50)
## Draw silhouette curve
GeneDropoutRatePlot(sce)

}
\keyword{average}
\keyword{curve}
\keyword{dropout}
\keyword{expression}
\keyword{gene}
\keyword{non-zero}
\keyword{rate}
