% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{GeneDropoutRatePlot}
\alias{GeneDropoutRatePlot}
\alias{GeneDropoutRatePlot.SingleCellExperiment}
\alias{GeneDropoutRatePlot,SingleCellExperiment-method}
\title{Dropout rate versus average non-zero expression}
\usage{
GeneDropoutRatePlot.SingleCellExperiment(
  object,
  genes,
  return.plot,
  clusters,
  clustering.type
)

\S4method{GeneDropoutRatePlot}{SingleCellExperiment}(
  object,
  genes = NULL,
  return.plot = FALSE,
  clusters = NULL,
  clustering.type = "manual"
)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{genes}{a character vector of the genes to visualize over
the dropout rate curve}

\item{return.plot}{return.plot whether to return the ggplot2 object or
just draw it. Default is \code{FALSE}.}

\item{clusters}{use data from these clusters only.
Default is \code{NULL}.}

\item{clustering.type}{"manual" or "optimal". "manual" refers to the
clustering formed using the "SelectKClusters" function and "optimal"
to the clustering using the "CalculateSilhouetteInformation" function.
Default is \code{manual}.}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
The GeneDropoutRatePlot function enables visualizion of dropout rates
(fraction of cells expressing a gene) against the average non-zero
expression values. This function can aid the user to determinine if
a gene is differentially expressed or not, since differentially expressed
genes are likely to be above the expected curve. Cells can be filtered based
on clustering before calculating the values.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(logcounts = pbmc3k_500))
sce <- PrepareILoReg(sce)
## These settings are used just to accelerate the example, use the defaults.
sce <- RunParallelICP(sce,L=2,threads=1,C=0.1,r=1,k=5) # Use L=200
sce <- RunPCA(sce,p=5)
sce <- HierarchicalClustering(sce)
sce <- SelectKClusters(sce,K=5)
gene_markers <- FindAllGeneMarkers(sce,log2fc.threshold = 0.5,min.pct = 0.5)
top10_log2FC <- SelectTopGenes(gene_markers,top.N=10,
criterion.type="log2FC",reverse=FALSE)
GeneHeatmap(sce,clustering.type = "manual",
 gene.markers = top10_log2FC)
## Dropout curve for cells of clusters 1 and 2. Mark and name three genes.
GeneDropoutRatePlot(sce, c("CD3D","CD14","CD79A"),clusters=1,2)

}
\keyword{average}
\keyword{curve}
\keyword{dropout}
\keyword{expression}
\keyword{gene}
\keyword{non-zero}
\keyword{rate}
