% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{GeneHeatmap}
\alias{GeneHeatmap}
\title{Heatmap visualization of the gene markers identified by FindAllGeneMarkers}
\usage{
\S4method{GeneHeatmap}{iloreg}(iloreg.object = NULL,
  clustering.type = "", gene.marker.data.frame = data.frame())
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{clustering.type}{"manual" or "optimal". "manual" refers to the clustering formed using the "SelectKClusters" function and "optimal" to the clustering using the "CalculateSilhouetteInformation" function. (default "manual")}

\item{gene.marker.data.frame}{a data frame of the gene markers generated by FindAllGeneMarkers function. To accelerate the drawing, filtering the dataframe by selecting e.g. top 10 genes is recommended.}
}
\value{
iloreg Object
}
\description{
The GeneHeatmap function enables drawing a heatmap of the gene markers identified by FindAllGeneMarkers, where the cell are grouped
by the clustering.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- HierarchicalClustering(iloreg_object)
iloreg_object <- SelectKClusters(iloreg_object,K=15)
gene_markers <- FindAllGeneMarkers(iloreg_object,
                                  clustering.type = "manual",
                                  test = "wilcox",
                                  logfc.threshold = 0.25,
                                  min.pct = 0.1,
                                 min.diff.pct = NULL,
                                 pseudocount.use = 1,
                                 min.cells.group = 3,
                                return.thresh = 0.01,
                                only.pos = FALSE,max.cells.per.cluster = NULL)
library(dplyr)
## Select top 10 genes per cluster by log2 fold-change
gene_markers \%>\% group_by(cluster) \%>\% top_n(10, log2FC) -> top10
gene_markers \%>\% group_by(cluster) \%>\% top_n(1, log2FC) -> top1

GeneHeatmap(iloreg_object,clustering.type = "manual",gene.marker.data.frame = top10)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
