% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{GeneScatterPlot}
\alias{GeneScatterPlot}
\title{Visualize gene expression over nonlinear dimensionality reduction}
\usage{
\S4method{GeneScatterPlot}{iloreg}(iloreg.object = NULL, genes = "",
  return.plot = FALSE, dim.reduction.type = "tsne", point.size = 0.7,
  title = "")
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{genes}{a character vector of the genes to be visualized}

\item{return.plot}{whether to return the ggplot2 object or just draw it (default \code{FALSE})}

\item{dim.reduction.type}{"tsne" or "umap" (default "tsne")}

\item{point.size}{point size (default 0.7)}

\item{title}{text to write above the plot}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
GeneScatterPlot enables visualizing gene expression of a gene over nonlinear dimensionality reduction with t-SNE or UMAP.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- RunTSNE(iloreg_object,perplexity=30)
GeneScatterPlot(iloreg_object,c("CCR7","CD3D","S100A4"),dim.reduction.type = "tsne",point.size=0.7)
iloreg_object <- RunUMAP(iloreg_object,perplexity=30)
GeneScatterPlot(iloreg_object,c("CCR7","CD3D","S100A4"),dim.reduction.type = "umap",point.size=0.7)
}
\keyword{gene}
\keyword{plot}
\keyword{scatter}
\keyword{visualization}
