% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{GeneScatterPlot}
\alias{GeneScatterPlot}
\title{Iterative logistic regression (ILoReg) consensus clustering}
\usage{
\S4method{GeneScatterPlot}{iloreg}(iloreg.object = NULL, genes = "",
  return.plot = FALSE, dim.reduction.type = "", point.size = 0.7)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{genes}{object of class 'iloreg'}

\item{return.plot}{object of class 'iloreg'}

\item{dim.reduction.type}{object of class 'iloreg'}

\item{point.size}{object of class 'iloreg'}
}
\value{
iloreg Object
}
\description{
Plot histogram of the ARIs.
}
\details{
populates a Boolean matrix with the same dimension as nData.
The value is \code{TRUE} for an entry if it
is a dropout candidate; otherwise the value is \code{FALSE}.
}
\examples{
a <- c(0,1,2)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
