% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{HierarchicalClustering}
\alias{HierarchicalClustering}
\alias{HierarchicalClustering.SingleCellExperiment}
\alias{HierarchicalClustering,SingleCellExperiment-method}
\title{Hierarchical clustering using the Ward's method}
\usage{
HierarchicalClustering.SingleCellExperiment(object)

\S4method{HierarchicalClustering}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Perform Hierarchical clustering using the Ward's method.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)

}
\keyword{clustering}
\keyword{hierarchical}
\keyword{ward}
