% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{pbmc3k_500}
\alias{pbmc3k_500}
\title{A toy dataset with 500 cells downsampled from the pbmc3k dataset.}
\format{
pbmc3k_500, dgCMatrix object
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression}
}
\usage{
data(pbmc3k_500)
}
\description{
The preprocessing was done using Cell Ranger v2.2.0 and
the GRCh38.p12 human reference genome.
The Normalization was done using the LogNormalize method of Seurat v3 R package.
The sampling was done using the sample() function without replacement
and set.seed(1) as initialization.
}
\examples{
data(pbmc3k_500)
}
\keyword{datasets}
