% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{PCAElbowPlot}
\alias{PCAElbowPlot}
\title{Elbow plot of the standard deviations of the principal components}
\usage{
\S4method{PCAElbowPlot}{iloreg}(iloreg.object = NULL,
  return.plot = FALSE)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{return.plot}{logical indicating if the ggplot2 object should be returned (default FALSE)}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
Draw an elbow plot of the standard deviations of the principal components to deduce an appropriate value for p.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
}
\keyword{PCA}
\keyword{elbow}
\keyword{plot}
