% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{PCAElbowPlot}
\alias{PCAElbowPlot}
\alias{PCAElbowPlot.SingleCellExperiment}
\alias{PCAElbowPlot,SingleCellExperiment-method}
\title{Elbow plot of the standard deviations of the principal components}
\usage{
PCAElbowPlot.SingleCellExperiment(object, return.plot)

\S4method{PCAElbowPlot}{SingleCellExperiment}(object, return.plot = FALSE)
}
\arguments{
\item{object}{object of class 'iloreg'}

\item{return.plot}{logical indicating if the ggplot2 object
should be returned (default FALSE)}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
Draw an elbow plot of the standard deviations of the principal components
to deduce an appropriate value for p.
}
\keyword{PCA}
\keyword{elbow}
\keyword{plot}
