% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RenameAllClusters}
\alias{RenameAllClusters}
\alias{RenameAllClusters.SingleCellExperiment}
\alias{RenameAllClusters,SingleCellExperiment-method}
\title{Renaming all clusters at once}
\usage{
RenameAllClusters.SingleCellExperiment(object, new.cluster.names)

\S4method{RenameAllClusters}{SingleCellExperiment}(object, new.cluster.names = "")
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{new.cluster.names}{object of class 'iloreg'}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
RenameAllClusters function enables renaming all cluster at once.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce, K.start = 2, K.end = 50)
## Draw silhouette curve
RenameAllClusters(sce)

}
\keyword{all}
\keyword{clusters}
\keyword{rename}
