% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{RenameAllClusters}
\alias{RenameAllClusters}
\title{Iterative logistic regression (ILoReg) consensus clustering}
\usage{
\S4method{RenameAllClusters}{iloreg}(iloreg.object = NULL,
  new.cluster.names = "")
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{new.cluster.names}{object of class 'iloreg'}
}
\value{
iloreg Object
}
\description{
Plot histogram of the ARIs.
}
\details{
populates a Boolean matrix with the same dimension as nData.
The value is \code{TRUE} for an entry if it
is a dropout candidate; otherwise the value is \code{FALSE}.
}
\examples{
a <- c(0,1,2)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
