% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\name{RenameCluster}
\alias{RenameCluster}
\title{Renaming one cluster}
\usage{
\S4method{RenameCluster}{iloreg}(
  iloreg.object = NULL,
  old.cluster.name = "",
  new.cluster.name = ""
)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{old.cluster.name}{object of class 'iloreg'}

\item{new.cluster.name}{object of class 'iloreg'}
}
\value{
iloreg object
}
\description{
RenameCluster function enables renaming a cluster
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- HierarchicalClustering(iloreg_object)
iloreg_object <- SelectKClusters(iloreg_object,K=15)
## Rename cluster 1 to "A"
iloreg_object <- RenameCluster(iloreg_object,old.cluster.name=1,new.cluster.name="A")
}
\keyword{cluster}
\keyword{one}
\keyword{rename}
