% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RenameCluster}
\alias{RenameCluster}
\alias{RenameCluster.SingleCellExperiment}
\alias{RenameCluster,SingleCellExperiment-method}
\title{Renaming one cluster}
\usage{
RenameCluster.SingleCellExperiment(object, old.cluster.name, new.cluster.name)

\S4method{RenameCluster}{SingleCellExperiment}(object, old.cluster.name = "", new.cluster.name = "")
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{old.cluster.name}{a character variable denoting  the
old name of the cluster}

\item{new.cluster.name}{a character variable the
new name of the cluster}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
RenameCluster function enables renaming
a cluster in `clustering.manual` slot.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce, K.start = 2, K.end = 50)
## Draw silhouette curve
RenameCluster(sce)

}
\keyword{cluster}
\keyword{one}
\keyword{rename}
