% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{RunTSNE}
\alias{RunTSNE}
\title{Barnes-Hut implementation of t-Distributed Stochastic Neighbor Embedding (t-SNE)}
\usage{
\S4method{RunTSNE}{iloreg}(iloreg.object = NULL, perplexity = 30)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{perplexity}{perplexity of t-SNE}
}
\value{
iloreg object
}
\description{
Run nonlinear dimensionality reduction using t-SNE with the PCA-transformed consensus matrix as input.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- RunTSNE(iloreg_object)
}
\keyword{Barnes-Hut}
\keyword{Embedding}
\keyword{Neighbor}
\keyword{Stochastic}
\keyword{implementation}
\keyword{of}
\keyword{t-Distributed}
\keyword{t-SNE}
