% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{RunUMAP}
\alias{RunUMAP}
\title{Uniform Manifold Approximation and Projection (UMAP)}
\usage{
\S4method{RunUMAP}{iloreg}(iloreg.object = NULL)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}
}
\value{
iloreg object
}
\description{
Run nonlinear dimensionality reduction using UMAP with the PCA-transformed consensus matrix as input.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- RunUMAP(iloreg_object)
}
\keyword{Approximation}
\keyword{Manifold}
\keyword{Projection}
\keyword{UMAP}
\keyword{Uniform}
\keyword{and}
