% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\name{SelectKClusters}
\alias{SelectKClusters}
\title{Selecting K clusters from hierarchical clustering}
\usage{
\S4method{SelectKClusters}{iloreg}(iloreg.object = NULL, K = 10)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{K}{a positive integer denoting how many clusters to select}
}
\value{
iloreg Object
}
\description{
Selects K clusters from the dendrogram using the cutree function from the dendextend R package.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- HierarchicalClustering(iloreg_object)
iloreg_object <- SelectKClusters(iloreg_object,K=25)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
