% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{SelectKClusters}
\alias{SelectKClusters}
\alias{SelectKClusters.SingleCellExperiment}
\alias{SelectKClusters,SingleCellExperiment-method}
\title{Selecting K clusters from hierarchical clustering}
\usage{
SelectKClusters.SingleCellExperiment(object, K)

\S4method{SelectKClusters}{SingleCellExperiment}(object, K = NULL)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{K}{a positive integer denoting how many clusters to select}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Selects K clusters from the dendrogram.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce, K.start = 2, K.end = 50)
## Draw silhouette curve
SelectKClusters(sce)

}
\keyword{clusters}
\keyword{select}
