% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{SelectTopGenes}
\alias{SelectTopGenes}
\title{Select top or bottom N genes based on a selection criterion}
\usage{
SelectTopGenes(
  gene.markers = NULL,
  top.N = 10,
  criterion.type = "log2FC",
  reverse = FALSE
)
}
\arguments{
\item{gene.markers}{A data frame of the gene markers found by
FindAllGeneMarkers function.}

\item{top.N}{How many top or bottom genes to select. Default is \code{10}.}

\item{criterion.type}{Which criterion to use for selecting the genes.
Default is "log2FC".}

\item{reverse}{Whether to select bottom instead of top N genes.
Default is \code{FALSE}.}
}
\value{
an object of `data.frame` class
}
\description{
The SelectTopGenes function enables selecting top or bottom N genes based
on a criterion (e.g. log2FC or adj.p.value).
}
\keyword{N}
\keyword{bottom}
\keyword{genes}
\keyword{select}
\keyword{top}
