% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VisualizeQC}
\alias{VisualizeQC}
\alias{VisualizeQC.SingleCellExperiment}
\alias{VisualizeQC,SingleCellExperiment-method}
\title{Visualize the terminal projection accuracy, number of
epochs/iterations and the average pairwise ARI values.}
\usage{
VisualizeQC.SingleCellExperiment(object, return.plot)

\S4method{VisualizeQC}{SingleCellExperiment}(object, return.plot = FALSE)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{return.plot}{logical indicating if the ggplot2 object
should be returned. Default is \code{FALSE}.}
}
\value{
ggplot2 object if \code{return.plot=TRUE}
}
\description{
Draw violin plots of the terminal projection accuracy,
the number of epochs/iterations and the average pairwise ARI values.
}
\keyword{ARI}
\keyword{QC}
\keyword{accuracy}
\keyword{average}
\keyword{control}
\keyword{pairwise}
\keyword{projection}
\keyword{quality}
\keyword{terminal}
