% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{VisualizeQC}
\alias{VisualizeQC}
\title{Visualize the terminal projection accuracy, number of epochs/iterations and the average pairwise ARI values.}
\usage{
\S4method{VisualizeQC}{iloreg}(iloreg.object = NULL,
  return.plot = FALSE)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{return.plot}{logical indicating if the ggplot2 object should be returned (default FALSE)}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
Draw violin plots of the terminal projection accuracy, the number of epochs/iterations and the average pairwise ARI values.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
}
\keyword{ARI}
\keyword{QC}
\keyword{accuracy}
\keyword{average}
\keyword{control}
\keyword{pairwise}
\keyword{projection}
\keyword{quality}
\keyword{terminal}
