% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{VisualizeQC}
\alias{VisualizeQC}
\title{Iterative logistic regression (ILoReg) consensus clustering}
\usage{
\S4method{VisualizeQC}{iloreg}(iloreg.object = NULL,
  return.plot = FALSE)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{return.plot}{logical indicating if the ggplot2 object should be returned (default FALSE)}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
Plot histogram of the ARIs.
}
\details{
populates a Boolean matrix with the same dimension as nData.
The value is \code{TRUE} for an entry if it
is a dropout candidate; otherwise the value is \code{FALSE}.
}
\examples{
a <- c(0,1,2)
}
\keyword{ILoReg}
\keyword{clustering}
\keyword{consensus}
\keyword{iterative}
\keyword{logistic}
\keyword{regression}
