% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VlnPlot}
\alias{VlnPlot}
\alias{VlnPlot.SingleCellExperiment}
\alias{VlnPlot,SingleCellExperiment-method}
\title{Gene expression visualization using violin plots}
\usage{
VlnPlot.SingleCellExperiment(object, clustering.type, genes, return.plot)

\S4method{VlnPlot}{SingleCellExperiment}(object, clustering.type = "manual", genes = NULL, return.plot = FALSE)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{clustering.type}{"manual" or "optimal". "manual"
refers to the clustering formed using the "SelectKClusters" function
and "optimal" to the clustering formed using the
"CalculateSilhouetteInformation" function. Default is "manual".}

\item{genes}{a character vector denoting the gene names that are visualized}

\item{return.plot}{return.plot whether to return the ggplot2 object
or just draw it. Default is \code{FALSE}.}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
The VlnPlot function enables visualizing expression levels of a gene,
or multiple genes, across clusters using Violin plots.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Create a SingleCellExperiment object
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays=list(logcounts=data))
## Cleans the data in `logcounts` slot: removes duplicate
## and non-expressing genes.
library(ILoReg)
sce <- PreapareILoReg(sce)
## Run L ICP runs in parallel
sce <- RunParallelICP(sce)
## Visualize QC
VisualizeQC(sce)
## Run PCA
sce <- RunPCA(sce,p = 50,scale = FALSE)
## Cluster hierarchically using the Ward's agglomeration
sce <- HierarchicalClustering(sce)
## Estimate optimal K using silhouette method
sce <- CalculateSilhouetteInformation(sce, K.start = 2, K.end = 50)
## Draw silhouette curve
VlnPlot(sce)

}
\keyword{plot}
\keyword{violin}
