% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILoReg.R
\docType{methods}
\name{VlnPlot}
\alias{VlnPlot}
\title{Gene expression visualization using violin plots}
\usage{
\S4method{VlnPlot}{iloreg}(iloreg.object = NULL, clustering.type = "",
  genes = "", return.plot = FALSE)
}
\arguments{
\item{iloreg.object}{object of class 'iloreg'}

\item{clustering.type}{"manual" or "optimal". "manual" refers to the clustering formed using the "SelectKClusters" function and "optimal" to the clustering formed using the "CalculateSilhouetteInformation" function. (default "manual")}

\item{genes}{a character vector denoting the gene names that are visualized}

\item{return.plot}{return.plot whether to return the ggplot2 object or just draw it (default \code{FALSE})}
}
\value{
ggplot2 object if return.plot=TRUE
}
\description{
The VlnPlot function enables visualizing expression levels of a gene, or multiple genes, across clusters using Violin plots.
}
\examples{
## Load 10X Chromium data and LogNormalized them.
raw_data <- Seurat::Read10x("~/../Downloads/10x_data/pbmc3k/filtered_gene_bc_matrices/hg19/")
data <- Seurat::LogNormalize(raw_data)
## Initialize the iloreg object with a normalized gene expression dataset.
iloreg_object <- CreateILoRegObject(normalized.data=data)
iloreg_object <- RunParallelICP(iloreg_object,threads = 12,seed=1)
VisualizeQC(iloreg_object)
iloreg_object <- RunPCA(iloreg_object,p = 50,scale = FALSE)
PCAElbowPlot(iloreg_object,return.plot=FALSE)
iloreg_object <- RunTSNE(iloreg_object,perplexity=30)
iloreg_object <- HierarchicalClustering(iloreg_object)
iloreg_object <- SelectKClusters(iloreg_object,K=15)
VlnPlot(iloreg_object,genes = c("CD3D","CD8B","CD79A"))
}
\keyword{plot}
\keyword{violin}
